+++ src/client/js/clear.js
@@ -1,0 +1,11 @@
+/*
+ * pwix:accounts-ui/src/client/js/clear.js
+ */
+
+/**
+ * @summary Clears the current panel
+ * @locus Client
+ */
+AccountsUI.clearPanel = function(){
+    console.debug( 'clearing panel' );
+};