+++ src/client/js/index.js
@@ -4,6 +4,7 @@
 import '../../common/js/index.js';
 
 import './accounts_base.js';
+import './clear.js';
 import './defaults.js';
 import './functions.js';
 import './ready.js';